Move event to non legacy folder
