+++ vsphere/datadog_checks/vsphere/api.py
@@ -12,7 +12,7 @@ from pyVmomi import vim, vmodl
 from datadog_checks.base.log import CheckLoggingAdapter
 from datadog_checks.vsphere.config import VSphereConfig
 from datadog_checks.vsphere.constants import ALL_RESOURCES, MAX_QUERY_METRICS_OPTION, UNLIMITED_HIST_METRICS_PER_QUERY
-from datadog_checks.vsphere.legacy.event import ALLOWED_EVENTS
+from datadog_checks.vsphere.event import ALLOWED_EVENTS
 from datadog_checks.vsphere.types import InfrastructureData
 
 CallableT = TypeVar('CallableT', bound=Callable)
