+++ vsphere/datadog_checks/vsphere/constants.py
@@ -3,6 +3,8 @@
 # Licensed under Simplified BSD License (see LICENSE)
 from pyVmomi import vim
 
+SOURCE_TYPE = 'vsphere'
+
 ALLOWED_FILTER_TYPES = ['whitelist', 'blacklist']
 ALLOWED_FILTER_PROPERTIES = ['name', 'inventory_path', 'tag']
 EXTRA_FILTER_PROPERTIES_FOR_VMS = ['hostname', 'guest_hostname']
