+++ vsphere/datadog_checks/vsphere/legacy/vsphere_legacy.py
@@ -25,10 +25,10 @@ from datadog_checks.base.checks.libs.vmware.all_metrics import ALL_METRICS
 from datadog_checks.base.checks.libs.vmware.basic_metrics import BASIC_METRICS
 from datadog_checks.base.config import is_affirmative
 
+from ..event import VSphereEvent
 from .cache_config import CacheConfig
 from .common import REALTIME_RESOURCES, SOURCE_TYPE
 from .errors import BadConfigError, ConnectionError
-from .event import VSphereEvent
 from .metadata_cache import MetadataCache, MetadataNotFoundError
 from .mor_cache import MorCache, MorNotFoundError
 from .objects_queue import ObjectsQueue
